[master] freddominette

+++ libft/ft_printf/statemachine/statemachine.c
@@ -6,7 +6,7 @@
 /*   By: fkante <fkante@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/26 17:37:23 by fkante            #+#    #+#             */
-/*   Updated: 2019/09/20 11:50:54 by amartino         ###   ########.fr       */
+/*   Updated: 2020/05/11 22:50:00 by francis          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,5 +22,5 @@ ssize_t		parser(t_state_machine *ptf, char *input, va_list *arg_pf)
 		if ((scale = parser[ptf->state](ptf, input, arg_pf)) != FAILURE)
 			input += scale;
 	}
-	return (ptf->state == ST_ERROR ? FAILURE : vct_len(ptf->output));
+	return (ptf->state == ST_ERROR ? FAILURE : (ssize_t)vct_len(ptf->output));
 }
